refactor(config): extract harness posture types into a harness module (#3311)
Third behavior-preserving slice of the config module split. Moves the harness
posture/profile domain -- HarnessPostureKind, HarnessCompactionStrategy,
HarnessToolSurface, HarnessSafetyPosture, HarnessPosture (+ Default/impl), and
HarnessProfile (+ matches_route) -- verbatim out of lib.rs into a focused
`harness` module, re-exported at the crate root so every existing path is
unchanged. The module references the crate-private route/wildcard matchers
(provider_routes_equal, wildcard_pattern_matches) via crate::; no values or
logic changed.

lib.rs shrinks ~143 lines (~4.24k now), separating the agent-posture domain
from the rest of the config schema per #3311 / #2608.

Verified: cargo fmt --check, cargo clippy --all-targets,
cargo test -p codewhale-config (251 passed), scripts/check-provider-registry.py
passes.
